Restore mstest and remove license url
