+++ .travis.yml
@@ -7,11 +7,10 @@ dotnet: 2.1.503
 install:
 - export FrameworkPathOverride=$(dirname $(which mono))/../lib/mono/4.5/
 - dotnet restore bill-payments-dotnet-sdk.sln
-- nuget install NUnit.Runners -OutputDirectory testrunner
 - if [[ $TRAVIS_PULL_REQUEST = 'false' && $TRAVIS_BRANCH = 'master' ]]; then openssl aes-256-cbc -K $encrypted_35f90653928635e2_key -iv $encrypted_35f90653928635e2_iv -in Qiwi.BillPayments.Tests/App.config.enc -out Qiwi.BillPayments.Tests/App.config -d; fi
 script:
 - msbuild /t:Rebuild /p:Configuration=Release bill-payments-dotnet-sdk.sln
-- mono ./testrunner/NUnit.ConsoleRunner.*/tools/nunit3-console.exe ./Qiwi.BillPayments.Tests/bin/Release/netcoreapp2.0/Qiwi.BillPayments.Tests.dll
+- dotnet test bill-payments-dotnet-sdk.sln
 deploy:
   skip_cleanup: true
   provider: script
