+++ Qiwi.BillPayments.Tests/Client/BillPaymentClientTest.Api.cs
@@ -1,8 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Web;
-using NUnit.Framework;
-using NUnit.Framework.Internal;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Qiwi.BillPayments.Client;
 using Qiwi.BillPayments.Json;
 using Qiwi.BillPayments.Json.Newtonsoft;
@@ -12,8 +11,8 @@ using Qiwi.BillPayments.Model.Out;
 
 namespace Qiwi.BillPayments.Tests.Client
 {
-    [TestFixture]
-    public class BillPaymentClientApiTest : BillPaymentClientTestCase
+    [TestClass]
+    public partial class BillPaymentClientTest
     {
         private class ApiPresets
         {
@@ -21,37 +20,42 @@ namespace Qiwi.BillPayments.Tests.Client
             public List<RefundResponse> Refunds { get; set; }
         }
         
-        private static readonly IReadOnlyDictionary<BillPaymentsClient, ApiPresets> ApiDataRows
-            = new Dictionary<BillPaymentsClient, ApiPresets>
-            {
+        private static IReadOnlyDictionary<BillPaymentsClient, ApiPresets> ApiDataRows
+            => (Dictionary<BillPaymentsClient, ApiPresets>) _testContext.Properties["apiDataRows"];
+        
+        private static void ClassInitialize_Api()
+        {
+            _testContext.Properties.Add(
+                "apiDataRows",
+                new Dictionary<BillPaymentsClient, ApiPresets>
                 {
-                    BillPaymentsClientFactory.create(Config.MerchantSecretKey),
-                    new ApiPresets
                     {
-                        Bills = new List<BillResponse>(),
-                        Refunds = new List<RefundResponse>()
-                    }
-                },
-                {
-                    BillPaymentsClientFactory.create(
-                        Config.MerchantSecretKey,
-                        null,
-                        ObjectMapperFactory.create<NewtonsoftMapper>()
-                    ),
-                    new ApiPresets
+                        BillPaymentsClientFactory.create(Config.MerchantSecretKey),
+                        new ApiPresets
+                        {
+                            Bills = new List<BillResponse>(),
+                            Refunds = new List<RefundResponse>()
+                        }
+                    },
                     {
-                        Bills = new List<BillResponse>(),
-                        Refunds = new List<RefundResponse>()
+                        BillPaymentsClientFactory.create(
+                            Config.MerchantSecretKey,
+                            null,
+                            ObjectMapperFactory.create<NewtonsoftMapper>()
+                        ),
+                        new ApiPresets
+                        {
+                            Bills = new List<BillResponse>(),
+                            Refunds = new List<RefundResponse>()
+                        }
                     }
                 }
-            };
+            );
+        }
         
-        [Test]
-        [Sequential]
-        public void TestCreatePaymentForm(
-            [Values("200.345")] string value,
-            [Values("http://test.ru/")] string successUrl
-        )
+        [TestMethod]
+        [DataRow("200.345", "http://test.ru/")]
+        public void TestCreatePaymentForm(string value, string successUrl)
         {
             // Prepare
             foreach (var (client, _) in ApiDataRows)
@@ -81,17 +85,17 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Sequential]
-        [Order(1)]
+        [TestMethod]
+        [Priority(1)]
+        [DataRow("200.345", "RUB", "test", "test@test.ru", "user uid on your side", "79999999999", "http://test.ru/")]
         public void TestCreateBill_Api(
-            [Values("200.345")] string value,
-            [Values("RUB")] string currency,
-            [Values("test")] string comment,
-            [Values("test@test.ru")] string email,
-            [Values("user uid on your side")] string account,
-            [Values("79999999999")] string phone,
-            [Values("http://test.ru/")] string successUrl
+            string value,
+            string currency,
+            string comment,
+            string email,
+            string account,
+            string phone,
+            string successUrl
         )
         {
             Config.Required();
@@ -117,8 +121,8 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(2)]
+        [TestMethod]
+        [Priority(2)]
         public void TestGetBillInfo_Api()
         {
             Config.Required();
@@ -136,8 +140,8 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(3)]
+        [TestMethod]
+        [Priority(3)]
         public void TestCancelBill_Api()
         {
             Config.Required();
@@ -155,13 +159,10 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(1)]
-        [Sequential]
-        public void TestRefundBill_Api(
-            [Values("0.01")] string amount,
-            [Values("RUB")] string currency
-        )
+        [TestMethod]
+        [Priority(1)]
+        [DataRow("0.01", "RUB")]
+        public void TestRefundBill_Api(string amount, string currency)
         {
             Config.Required();
             // Prepare
@@ -181,8 +182,8 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(2)]
+        [TestMethod]
+        [Priority(2)]
         public void TestGetRefundInfo_Api()
         {
             Config.Required();
