+++ Qiwi.BillPayments.Tests/Client/BillPaymentClientTest.Fake.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
-using NUnit.Framework;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Qiwi.BillPayments.Client;
 using Qiwi.BillPayments.Json;
 using Qiwi.BillPayments.Json.Newtonsoft;
@@ -10,53 +10,58 @@ using Qiwi.BillPayments.Model;
 using Qiwi.BillPayments.Model.In;
 using Qiwi.BillPayments.Model.Out;
 using Qiwi.BillPayments.Tests.Web;
+using Qiwi.BillPayments.Utils;
 
 namespace Qiwi.BillPayments.Tests.Client
 {
-    [TestFixture]
-    public class BillPaymentClientFakeTest : BillPaymentClientTestCase
+    public partial class BillPaymentClientTest
     {
-        private class FakeApiPresets
+        private class FakeApiPresets : ApiPresets
         {
-            public List<BillResponse> Bills { get; set; }
-            public List<RefundResponse> Refunds { get; set; }
             public FakeClient Client { get; set; }
         }
         
-        private static readonly IReadOnlyDictionary<BillPaymentsClient, FakeApiPresets> FakeDataRows
-            = new Dictionary<BillPaymentsClient, FakeApiPresets>(
-            new[]
-                {
-                    new FakeClient(ObjectMapperFactory.create()),
-                    new FakeClient(ObjectMapperFactory.create<NewtonsoftMapper>())
-                }
-                .Select(fakeClient => new KeyValuePair<BillPaymentsClient, FakeApiPresets>(
-                    BillPaymentsClientFactory.create(Config.MerchantSecretKey, fakeClient, fakeClient.ObjectMapper),
-                    new FakeApiPresets
-                    {
-                        Client = fakeClient,
-                        Bills = new List<BillResponse>(),
-                        Refunds = new List<RefundResponse>()
-                    }
-                ))
-            );
+        private readonly IReadOnlyDictionary<BillPaymentsClient, FakeApiPresets> fakeDataRows
+            = (IReadOnlyDictionary<BillPaymentsClient, FakeApiPresets>) _testContext.Properties["fakeDataRows"];
         
-        [Test]
-        [Order(1)]
-        [Sequential]
+        private static void ClassInitialize_Fake()
+        {
+            _testContext.Properties.Add(
+                "fakeDataRows",
+                new Dictionary<BillPaymentsClient, FakeApiPresets>(
+                new[]
+                    {
+                        new FakeClient(ObjectMapperFactory.create()),
+                        new FakeClient(ObjectMapperFactory.create<NewtonsoftMapper>())
+                    }
+                    .Select(fakeClient => new KeyValuePair<BillPaymentsClient, FakeApiPresets>(
+                        BillPaymentsClientFactory.create(Config.MerchantSecretKey, fakeClient, fakeClient.ObjectMapper),
+                        new FakeApiPresets
+                        {
+                            Client = fakeClient,
+                            Bills = new List<BillResponse>(),
+                            Refunds = new List<RefundResponse>()
+                        }
+                    ))
+                )
+            );
+        }
+        
+        [TestMethod]
+        [Priority(1)]
+        [DataRow("200.345", "RUB", "test", "test@test.ru", "user uid on your side", "79999999999", "http://test.ru/")]
         public void TestCreateBill_Fake(
-            [Values("200.345")] string value,
-            [Values("RUB")] string currency,
-            [Values("test")] string comment,
-            [Values("test@test.ru")] string email,
-            [Values("user uid on your side")] string account,
-            [Values("79999999999")] string phone,
-            [Values("http://test.ru/")] string successUrl,
-            [Values("https://oplata.qiwi.com/form/?invoice_uid=bb773791-9bd9-42c1-b8fc-3358cd108422")] string payUri
+            string value,
+            string currency,
+            string comment,
+            string email,
+            string account,
+            string phone,
+            string successUrl
         )
         {
             // Prepare
-            foreach (var (client, presets) in FakeDataRows)
+            foreach (var (client, presets) in fakeDataRows)
             {
                 var fingerprint = client.getFingerprint();
                 PrepareBill(
@@ -76,6 +81,7 @@ namespace Qiwi.BillPayments.Tests.Client
                     Comment = createBillInfo.Comment,
                     ExpirationDateTime = createBillInfo.ExpirationDateTime,
                     Customer = createBillInfo.Customer,
+                    PayUrl = PayUri,
                     Status = new ResponseStatus
                     {
                         ValueEnum = BillStatusEnum.Waiting,
@@ -85,8 +91,7 @@ namespace Qiwi.BillPayments.Tests.Client
                     {
                         ApiClient = fingerprint.getClientName(),
                         ApiClientVersion = fingerprint.getClientVersion()
-                    },
-                    PayUrl = new Uri(payUri)
+                    }
                 };
                 ResetClient(presets.Client, billResponse);
                 // Test
@@ -99,12 +104,12 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(2)]
+        [TestMethod]
+        [Priority(2)]
         public void TestGetBillInfo_Fake()
         {
             // Prepare
-            foreach (var (client, presets) in FakeDataRows)
+            foreach (var (client, presets) in fakeDataRows)
             {
                 Assert.AreNotSame(0, presets.Bills.Count, "Bill will be create");
                 foreach (var bill in presets.Bills)
@@ -121,12 +126,12 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(3)]
+        [TestMethod]
+        [Priority(3)]
         public void TestCancelBill_Fake()
         {
             // Prepare
-            foreach (var (client, presets) in FakeDataRows)
+            foreach (var (client, presets) in fakeDataRows)
             {
                 Assert.AreNotSame(0, presets.Bills.Count, "Bill will be create");
                 foreach (var bill in presets.Bills)
@@ -162,16 +167,13 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(1)]
-        [Sequential]
-        public void TestRefundBill_Fake(
-            [Values("0.01")] string amount,
-            [Values("RUB")] string currency
-        )
+        [TestMethod]
+        [Priority(1)]
+        [DataRow("0.01", "RUB")]
+        public void TestRefundBill_Fake(string amount, string currency)
         {
             // Prepare
-            foreach (var (client, presets) in FakeDataRows)
+            foreach (var (client, presets) in fakeDataRows)
             {
                 PrepareRefund(
                     amount,
@@ -197,12 +199,12 @@ namespace Qiwi.BillPayments.Tests.Client
             }
         }
         
-        [Test]
-        [Order(2)]
+        [TestMethod]
+        [Priority(2)]
         public void TestGetRefundInfo_Fake()
         {
             // Prepare
-            foreach (var (client, presets) in FakeDataRows)
+            foreach (var (client, presets) in fakeDataRows)
             {
                 Assert.AreNotSame(0, presets.Bills.Count, "Refund will be create");
                 foreach (var refund in presets.Refunds)
