+++ Qiwi.BillPayments.Tests/Qiwi.BillPayments.Tests.csproj
@@ -7,9 +7,9 @@
     </PropertyGroup>
     
     <ItemGroup>
-        <PackageReference Include="nunit" Version="3.10.1" />
-        <PackageReference Include="NUnit3TestAdapter" Version="3.10.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.8.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
+        <PackageReference Include="MSTest.TestAdapter" Version="1.3.2" />
+        <PackageReference Include="MSTest.TestFramework" Version="1.3.2" />
         <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.6.0-preview.18571.3" />
     </ItemGroup>
