+++ Qiwi.BillPayments.Tests/Utils/BillPaymentsUtilsTest.cs
@@ -1,46 +1,49 @@
 using System;
-using NUnit.Framework;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Qiwi.BillPayments.Model;
 using Qiwi.BillPayments.Utils;
 
 namespace Qiwi.BillPayments.Tests.Utils
 {
-    [TestFixture]
+    [TestClass]
     public class BillPaymentsUtilsTest
     {
-        [Test]
-        [Sequential]
-        public void TestFormatValue(
-            [Values("200.345")] string iValue,
-            [Values("200.34")] string oValue
-        )
+        [TestMethod]
+        [DataRow("200.345", "200.34")]
+        public void TestFormatValue(string iValue, string oValue)
         {
             var value = BillPaymentsUtils.formatValue(iValue);
             Assert.AreEqual(oValue, value, "Equal format value");
         }
 
-        [Test]
-        [Sequential]
-        public void TestGetTimeoutDate(
-            [Values(45,   1)]    int offset,
-            [Values(null, 1.0d)] double? days
-        )
+        [TestMethod]
+        [DataRow(45, null)]
+        [DataRow(1, 1.0)]
+        public void TestGetTimeoutDate(int offset, double? days)
         {
             var value = BillPaymentsUtils.getTimeoutDate(days);
             Assert.IsTrue(DateTime.Now < value, "Timeout date in future");
             Assert.AreEqual(offset, (value - DateTime.Now).Days + 1, "Timeout date offset");   
         }
         
-        [Test]
-        [Sequential]
+        [TestMethod]
+        [DataRow(
+            "test-merchant-secret-for-signature-check",
+            "07e0ebb10916d97760c196034105d010607a6c6b7d72bfa1c3451448ac484a3b",
+            "test",
+            "test_bill",
+            "1.00",
+            "RUB",
+            "PAID"
+        )]
         public void TestCheckNotificationSignature(
-            [Values("test-merchant-secret-for-signature-check")] string merchantSecret,
-            [Values("07e0ebb10916d97760c196034105d010607a6c6b7d72bfa1c3451448ac484a3b")] string signature,
-            [Values("test")] string siteId,
-            [Values("test_bill")] string billId,
-            [Values("1.00")] string value,
-            [Values("RUB")] string currency,
-            [Values("PAID")] string status
+            string merchantSecret,
+            string signature,
+            string siteId,
+            string billId,
+            string value,
+            string currency,
+            string status
         )
         {
             var notification = new Notification
